# [Merged by Bors] - chore(algebra/category/Module/monoidal): Speed up the elaboration